// 3.Kickstart/ethereum/src/generatedTypes/CampaignFactory.ts
/* Autogenerated file. Do not edit manually. */
/* tslint:disable */
/* eslint-disable */

import BN from "bn.js";
import { ContractOptions } from "web3-eth-contract";
import { EventLog } from "web3-core";
import { EventEmitter } from "events";
import {
  Callback,
  PayableTransactionObject,
  NonPayableTransactionObject,
  BlockType,
  ContractEventLog,
  BaseContract,
} from "./types";

export interface EventOptions {
  filter?: object;
  fromBlock?: BlockType;
  topics?: string[];
}

export interface CampaignFactory extends BaseContract {
  constructor(
    jsonInterface: any[],
    address?: string,
    options?: ContractOptions
  ): CampaignFactory;
  clone(): CampaignFactory;
  methods: {
    createCampaign(
      minimum: number | string | BN
    ): NonPayableTransactionObject<void>;

    deployedCampaigns(
      arg0: number | string | BN
    ): NonPayableTransactionObject<string>;

    getDeployedCampaigns(): NonPayableTransactionObject<string[]>;
  };
  events: {
    allEvents(options?: EventOptions, cb?: Callback<EventLog>): EventEmitter;
  };
}